update: Ensure we invalidate the database to find new updates
Previously we retained a cached copy of the database, so despite actually
running the update on the index, we were still using an outdated version,
so we'd always hash_match on the first update we ever found. In most cases
this means the user sees the update notification once upon login, and
never again.

With this change we dynamically discover and report the updates as soon
as we encounter *new* updates. The decision was also made to switch to
sha256 to prevent hash collision issues.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>