updates_view: Support clickable maniphest URLs in the changelog
Frequently Solus updates will mention an associated Maniphest Task on
the Solus Phabricator instance in relation to the commit. When encountering
anything that looks like a Maniphest ID, i.e. TXXXX, we'll replace it with
a href that the user can click on for further details.

This change also fixes some buggy assumptions in the CVE changelog parsing
to ensure that the remnants of the changelog line aren't also butchered.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>