details: Begin a revamp of the details view
This change wraps up the details portion into a GtkStack with a new
switcher at the top, splitting the main UI into "Details" and "Changelog".
This will allow us to add a changelog view to every package in the Software
Center, without first having to wait for the update to see it from one
place.

For now we stick the meta information into a grid on the far right at the
window, with the idea being we'll fill this out and split it up into
multiple sections, and add a screenshot section above the details/changelog
switcher.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>