fetcher: Implement all the async core logic to load images
We may end up loading the local images instead of downloading, but we
still do this in an async fashion, to do the heavy GdkPixbuf loading
off the main thread. Once we have an image, whether we downloaded it
or loaded it from a previous run, we'll emit a signal with the loaded
pixbuf, and pass it to the consumer (i.e. package details view).

Signed-off-by: Ikey Doherty <ikey@solus-project.com>