fetcher: Use a separate load queue to the fetch queue
To ensure we never hit a local denial whereby we can't load an image
already fetched, due to waiting to fetch more, we use a separate load
queue to grab them.

Additionally, we stop using lambdas between the idles and threads because
the closures only contain a weak reference, and end up trashing the contents
of the signals causing false positives for things that actually did work.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>