search: Improve searches when using spaces
Typically uses might be looking for something like "gnome-multi-writer",
but type "gnome multi writer" as the search term. Given that we're using
a basic regex behind the scenes, help them out and allow a space character
to also match underscores and hyphens, which then also includes many more
of our packages in the results.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>