Initial experiment with the new 'Featured' view
This will eventually be provided by some JSON file or similar for us
to maintain on the server, but for now we'll just hardcode a list of
"featured" packages.

Now that the basic somewhat crumby framework is in, we can try to make it
look better and have it interactive.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>