categories: Load components in on a thread
This makes it feel like the SC is much faster, but obviously this now
introduces nasty race conditions. So to alleviate this we'll restore the
loading screen mechanism with the witty cruft in a follow up commit and
ensure we can't break the SC by rapidly switching between components.

I'd rather we present a complete component to users with a small loading
screen instead of resorting to cancellables and potential thread safety
issues.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>