Start fleshng out the eopkg dbus callbacks so we know the story
In a different vein to the old mechanism we're not *so* fussed about when
finished happens because we don't use async calls, we do it on a thread.
This just allows us to know when we need to go about rebuilding the indexes
for the plugin and ask the context to reload if necessary.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>