plugins: Add new Transaction type for the plugin API
We need a way to track the planned install/removal in a safe fashion,
and that needs to factor in any non-primary-op issues that may arise.
For example, the installation of gdm will drag in extra dependencies,
but it may *also* result in the removal of lightdm. Track everything
that is going to happen so we can also start telling users about stuff
we're going to remove (which we didn't do in the past.)

Signed-off-by: Ikey Doherty <ikey@solus-project.com>