plugins/eopkg: Always account for system.base updates
eopkg will forcibly update system.base ("mandatory updates") so we'll
actually need to know whats being done in that respect before we begin
our update, so we don't break the software center. It will also account
for the reverse dependency updates that result from mandatory updates
so we keep the system consistent.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>