Begin adding cross-talk in plugins for Ldm query of foreign items
We need our own plugin to talk back to the other plugins to get them
to find out the provider in relation to the name we have, thus it is
up to the modalias plugins to have the correct IDs and prefixes.

Long story short we'll now use this to flesh out the eopkg plugin to
return the kernel-appropriate providers for the requested provider package
name and link them from the LdmItem.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>