plugins/base: Add new notion of ItemLink for foreign items
This expands the API a little bit to accomodate LDM by introducing the
notion of an ItemLink. Effectively a Link is a reason for why two items
are now classed as "related".

In terms of ItemLink.ENHANCES, this allows soft recommendations in
the traditional sense of "this is not mandatory but would improve the
software", and will later be tied in for 32-bit recommendations.

For the ItemLink.PROVIDER link, this will be used in pure-virtual cases
where an Item by itself is not installable. We'll do away with the current
notion of "ItemStatus.META_HARDWARE" in a subsequent commit and instead
introduce a VIRTUAL status.

This will allow Devices to "provide" (via links) installable packages
directly in the Software Centre for linkage.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>