Fix Task and CVE links in changelogs/markdown
If tasks or CVEs are linked to directly in a phabricator diff, solus-sc tried to create a link inside of a link. Fix this by only checking for matches after a space. Also fix CVE links, which probably didn't work for quite some time because the base URI wasn't added.