Switch office.maths to use accessories-calculator by default.
Get the icon theme and pass it along to our ScComponentButton creation. This is used to perform icon theme specific changes, so we now will set the following icons for Breeze for the respective components:

- games.card: kpat
- games.puzzle: kblocks
- network.download: kget
- network.news: quiterss
- office.notes: knotes

Additionally we're now once again enforcing a set pixel size to ensure we properly downscale. There is still more components that need coverage, but those will occur in subsequent commits.