Deprecate elementary-icon-theme and packagekit.
packagekit has effectively been deprecated in Solus since March 2020. We have no plans on supporting or using it and we have packagekit disabled explicitly anywhere applicable.

elementary-icon-theme was deprecated as it was included in an era where we had a more laxed package inclusion policy. This theme is specific to elementary OS and in alignment with our inclusion policy of not allowing distribution-specific software and its lack of support for most applicables outside of elementary OS, it doesn't offer any value add.