Convert from appstream-glib to appstream
appstream-glib is under "heavy-maintenance", it also has some
incompatiblities with appstream metainfo generated by
`appstream-generator` as opposed to `appstream-builder`.

Appstream is newer, better supports the specification and is better
maintained.

The only change is functionality is that the thumbnail screenshot
size has doubled to 224px from 114px which slightly changes the UI.
This can be re-scaled down in future if desired.

Due to the change in APIs a helper `get_app_by_pkgname()` is introduced
to largely keep the same code structure.